Refactored view tracking

+++ qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/recording/ViewTrackerSlideOverview.java
@@ -1,3 +1,26 @@
+/*-
+ * #%L
+ * This file is part of QuPath.
+ * %%
+ * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
+ * Contact: IP Management (ipmanagement@qub.ac.uk)
+ * Copyright (C) 2018 - 2021 QuPath developers, The University of Edinburgh
+ * %%
+ * QuPath is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * QuPath is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License 
+ * along with QuPath.  If not, see <https://www.gnu.org/licenses/>.
+ * #L%
+ */
+
 package qupath.lib.gui.viewer.recording;
 
 import java.awt.Shape;
@@ -40,7 +63,7 @@ class ViewTrackerSlideOverview {
 		if (img == null)
 			return;
 
-		int preferredHeight = (int)(img.getHeight() * (double)(preferredWidth / (double)img.getWidth()));
+		int preferredHeight = (int)(img.getHeight() * (preferredWidth / (double)img.getWidth()));
 		var imgPreview = GuiTools.getScaledRGBInstance(img, preferredWidth, preferredHeight);
 		canvas.setWidth(imgPreview.getWidth());
 		canvas.setHeight(imgPreview.getHeight());
@@ -100,7 +123,7 @@ class ViewTrackerSlideOverview {
 	}
 	
 	private void drawImage(GraphicsContext g, BufferedImage imgToDraw) {
-		int preferredHeight = (int)(img.getHeight() * (double)(preferredWidth / (double)img.getWidth()));
+		int preferredHeight = (int)(img.getHeight() * (preferredWidth / (double)img.getWidth()));
 		var imgPreview = GuiTools.getScaledRGBInstance(imgToDraw, preferredWidth, preferredHeight);
 		g.drawImage(imgPreview, 0, 0);
 	}
